Select with optional checkbox or radio button groups
Added a "style" property to Select, which enables setting how the selection
is displayed and handled. Without setting "style" or setting it to "dropdown"
we get the original functionality with a Select dropdown.

Setting it to "checkbox" will show all select options at once with a checkbox
in front for selection.

If "style" is "radio," then all options are listed with a radio button in
front of each, allowing the selection of one item.

The specialty of the checkbox select is that it handles all value additions
on its own, e.g., values are not added as with other AddComponents one-by-one,
but the value coming from the checkbox select is the array of all current
selections. In case there may be other such inputs in the future, this is
generalized as the "handlesMultipleValues" feature. This value is set to
true implicitly by the ProfileManager for inputs that handle their array
values on their own and not one-by-one. This mechnism required just some
minimal additions to the CrateManager and render entity generation logic.